backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.3.1
django-filter==23.3
setuptools>=65.0.0
whitenoise==6.5.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
